+++ flexmeasures/api/v3_0/sensors.py
@@ -6,7 +6,7 @@ from flask_classful import FlaskView, route
 from flask_json import as_json
 from flask_security import auth_required
 import isodate
-from marshmallow import fields, validate, ValidationError
+from marshmallow import fields, ValidationError
 from marshmallow.validate import OneOf
 from rq.job import Job, NoSuchJobError
 from timely_beliefs import BeliefsDataFrame
@@ -34,12 +34,10 @@ from flexmeasures.data.models.user import Account
 from flexmeasures.data.models.time_series import Sensor
 from flexmeasures.data.queries.utils import simplify_index
 from flexmeasures.data.schemas.sensors import SensorSchema, SensorIdField
-from flexmeasures.data.schemas.times import AwareDateTimeField
-from flexmeasures.data.schemas.units import QuantityField
+from flexmeasures.data.schemas.times import AwareDateTimeField, PlanningDurationField
 from flexmeasures.data.schemas.scheduling import FlexContextSchema
 from flexmeasures.data.services.sensors import get_sensors
 from flexmeasures.data.services.scheduling import (
-    find_scheduler_class,
     create_scheduling_job,
     get_data_source_for_job,
 )
@@ -220,13 +218,12 @@ class SensorAPI(FlaskView):
                 data_key="start", format="iso", required=True
             ),
             "belief_time": AwareDateTimeField(format="iso", data_key="prior"),
+            "duration": PlanningDurationField(
+                load_default=PlanningDurationField.load_default
+            ),
             "flex_model": fields.Dict(data_key="flex-model"),
             "soc_sensor_id": fields.Str(data_key="soc-sensor", required=False),
-            "roundtrip_efficiency": QuantityField(
-                "%",
-                validate=validate.Range(min=0, max=1),
-                data_key="roundtrip-efficiency",
-            ),
+            "roundtrip_efficiency": fields.Str(data_key="roundtrip-efficiency"),
             "start_value": fields.Float(data_key="soc-at-start"),
             "soc_min": fields.Float(data_key="soc-min"),
             "soc_max": fields.Float(data_key="soc-max"),
@@ -262,6 +259,7 @@ class SensorAPI(FlaskView):
         self,
         sensor: Sensor,
         start_of_schedule: datetime,
+        duration: timedelta,
         belief_time: Optional[datetime] = None,
         start_value: Optional[float] = None,
         soc_min: Optional[float] = None,
@@ -298,9 +296,11 @@ class SensorAPI(FlaskView):
                   See https://github.com/FlexMeasures/flexmeasures/issues/485. Until then, it is possible to call this endpoint for one flexible endpoint at a time
                   (considering already scheduled sensors as inflexible).
 
-        The length of schedules is set by the config setting :ref:`planning_horizon_config`, defaulting to 12 hours.
-
-        .. todo:: add a schedule duration parameter, instead of always falling back to FLEXMEASURES_PLANNING_HORIZON
+        The length of the schedule can be set explicitly through the 'duration' field.
+        Otherwise, it is set by the config setting :ref:`planning_horizon_config`, which defaults to 48 hours.
+        If the flex-model contains targets that lie beyond the planning horizon, the length of the schedule is extended to accommodate them.
+        Finally, the schedule length is limited by :ref:`max_planning_horizon_config`, which defaults to 169 hours.
+        Targets that exceed the max planning horizon are not accepted.
 
         The appropriate algorithm is chosen by FlexMeasures (based on asset type).
         It's also possible to use custom schedulers and custom flexibility models, see :ref:`plugin_customization`.
@@ -323,8 +323,8 @@ class SensorAPI(FlaskView):
 
         **Example request B**
 
-        This message triggers a schedule for a storage asset, starting at 10.00am, at which the state of charge (soc) is 12.1 kWh,
-        with a target state of charge of 25 kWh at 4.00pm.
+        This message triggers a 24-hour schedule for a storage asset, starting at 10.00am,
+        at which the state of charge (soc) is 12.1 kWh, with a target state of charge of 25 kWh at 4.00pm.
         The minimum and maximum soc are set to 10 and 25 kWh, respectively.
         Roundtrip efficiency for use in scheduling is set to 98%.
         Aggregate consumption (of all devices within this EMS) should be priced by sensor 9,
@@ -337,6 +337,7 @@ class SensorAPI(FlaskView):
 
             {
                 "start": "2015-06-02T10:00:00+00:00",
+                "duration": "PT24H",
                 "flex-model": {
                     "soc-at-start": 12.1,
                     "soc-unit": "kWh",
@@ -409,8 +410,6 @@ class SensorAPI(FlaskView):
                 flex_model = {}
             if param is not None:
                 if param_name not in flex_model:
-                    if param_name == "roundtrip-efficiency" and type(param) != float:
-                        param = param.to(ur.Quantity("dimensionless")).magnitude  # type: ignore
                     flex_model[param_name] = param
                 found_fields["model"].append(param_name)
         # flex-context
@@ -447,9 +446,7 @@ class SensorAPI(FlaskView):
             )
         # -- end deprecation logic
 
-        end_of_schedule = start_of_schedule + current_app.config.get(  # type: ignore
-            "FLEXMEASURES_PLANNING_HORIZON"
-        )
+        end_of_schedule = start_of_schedule + duration
         scheduler_kwargs = dict(
             sensor=sensor,
             start=start_of_schedule,
@@ -461,20 +458,15 @@ class SensorAPI(FlaskView):
         )
 
         try:
-            # We create a scheduler, so the flex config is also checked and errors are returned here
-            scheduler = find_scheduler_class(sensor)(**scheduler_kwargs)
-            scheduler.deserialize_config()
+            job = create_scheduling_job(
+                **scheduler_kwargs,
+                enqueue=True,
+            )
         except ValidationError as err:
             return invalid_flex_config(err.messages)
         except ValueError as err:
             return invalid_flex_config(str(err))
 
-        job = create_scheduling_job(
-            **scheduler_kwargs,
-            enqueue=True,
-        )
-
-        scheduler.persist_flex_model()
         db.session.commit()
 
         response = dict(schedule=job.id)