+++ flexmeasures/data/models/planning/storage.py
@@ -205,6 +205,7 @@ class StorageScheduler(Scheduler):
         Deserialize storage flex model and the flex context against schemas.
         Before that, we fill in values from wider context, if possible.
         Mostly, we allow several fields to come from sensor attributes.
+        TODO: this work could maybe go to the schema as a pre-load hook (if we pass in the sensor to schema initialization)
 
         Note: Before we apply the flex config schemas, we need to use the flex config identifiers with hyphens,
               (this is how they are represented to outside, e.g. by the API), after deserialization
@@ -231,6 +232,12 @@ class StorageScheduler(Scheduler):
                 )
             else:
                 self.flex_model["soc-at-start"] = 0
+        # soc-unit
+        if "soc-unit" not in self.flex_model or self.flex_model["soc-unit"] is None:
+            if self.sensor.unit in ("MWh", "kWh"):
+                self.flex_model["soc-unit"] = self.sensor.unit
+            elif self.sensor.unit in ("MW", "kW"):
+                self.flex_model["soc-unit"] = self.sensor.unit + "h"
 
         # Check for round-trip efficiency
         if (
@@ -241,19 +248,38 @@ class StorageScheduler(Scheduler):
             self.flex_model["roundtrip-efficiency"] = self.sensor.get_attribute(
                 "roundtrip_efficiency", 1
             )
-        if (
-            self.flex_model["roundtrip-efficiency"] <= 0
-            or self.flex_model["roundtrip-efficiency"] > 1
-        ):
-            raise ValueError("roundtrip efficiency expected within the interval (0, 1]")
-
         self.ensure_soc_min_max()
 
         # Now it's time to check if our flex configurations holds up to schemas
-        self.flex_model = StorageFlexModelSchema().load(self.flex_model)
+        self.flex_model = StorageFlexModelSchema(self.start).load(self.flex_model)
         self.flex_context = FlexContextSchema().load(self.flex_context)
 
+        # Extend schedule period in case a target exceeds its end
+        self.possibly_extend_end()
+
         return self.flex_model
+
+    def possibly_extend_end(self):
+        """Extend schedule period in case a target exceeds its end.
+
+        The schedule's duration is possibly limited by the server config setting 'FLEXMEASURES_MAX_PLANNING_HORIZON'.
+
+        todo: when deserialize_flex_config becomes a single schema for the whole scheduler,
+              this function would become a class method with a @post_load decorator.
+        """
+        soc_targets = self.flex_model.get("soc_targets")
+        if soc_targets:
+            max_target_datetime = max(
+                [soc_target["datetime"] for soc_target in soc_targets]
+            )
+            if max_target_datetime > self.end:
+                max_server_horizon = current_app.config.get(
+                    "FLEXMEASURES_MAX_PLANNING_HORIZON"
+                )
+                if max_server_horizon:
+                    self.end = min(max_target_datetime, self.start + max_server_horizon)
+                else:
+                    self.end = max_target_datetime
 
     def get_min_max_targets(
         self, deserialized_names: bool = True
@@ -311,7 +337,7 @@ class StorageScheduler(Scheduler):
 
 
 def build_device_soc_targets(
-    targets: List[Dict[datetime, float]] | pd.Series,
+    targets: List[Dict[str, datetime | float]] | pd.Series,
     soc_at_start: float,
     start_of_schedule: datetime,
     end_of_schedule: datetime,
@@ -334,7 +360,7 @@ def build_device_soc_targets(
     TODO: this function could become the deserialization method of a new SOCTargetsSchema (targets, plural), which wraps SOCTargetSchema.
 
     """
-    if isinstance(targets, pd.Series):  # some teats prepare it this way
+    if isinstance(targets, pd.Series):  # some tests prepare it this way
         device_targets = targets
     else:
         device_targets = initialize_series(
@@ -351,9 +377,11 @@ def build_device_soc_targets(
                 device_targets.index.tzinfo
             )  # otherwise DST would be problematic
             if target_datetime > end_of_schedule:
-                raise ValueError(
-                    f'Target datetime exceeds {end_of_schedule}. Maximum scheduling horizon is {current_app.config.get("FLEXMEASURES_PLANNING_HORIZON")}.'
+                # Skip too-far-into-the-future target
+                current_app.logger.warning(
+                    f'Disregarding target datetime {target_datetime}, because it exceeds {end_of_schedule}. Maximum scheduling horizon is {current_app.config.get("FLEXMEASURES_MAX_PLANNING_HORIZON")}.'
                 )
+                continue
 
             device_targets.loc[target_datetime] = target_value