+++ flexmeasures/data/services/scheduling.py
@@ -5,11 +5,11 @@ import os
 import sys
 import importlib.util
 from importlib.abc import Loader
-from rq.job import Job
 
 from flask import current_app
 import click
 from rq import get_current_job
+from rq.job import Job
 import timely_beliefs as tb
 
 from flexmeasures.data import db
@@ -20,17 +20,9 @@ from flexmeasures.data.models.data_sources import DataSource
 from flexmeasures.data.utils import get_data_source, save_to_db
 from flexmeasures.utils.time_utils import server_now
 
-"""
-The life cycle of a scheduling job:
-1. A scheduling job is born in create_scheduling_job.
-2. It is run in make_schedule which writes results to the db.
-3. If an error occurs (and the worker is configured accordingly), handle_scheduling_exception comes in.
-   This might re-enqueue the job or try a different model (which creates a new job).
-"""
-
 
 def create_scheduling_job(
-    sensor: int | Sensor,
+    sensor: Sensor,
     job_id: str | None = None,
     enqueue: bool = True,
     **scheduler_kwargs,
@@ -42,14 +34,21 @@ def create_scheduling_job(
     That means one event leads to one job (i.e. actions are event driven).
 
     As a rule of thumb, keep arguments to the job simple, and deserializable.
+
+    The life cycle of a scheduling job:
+    1. A scheduling job is born here (in create_scheduling_job).
+    2. It is run in make_schedule which writes results to the db.
+    3. If an error occurs (and the worker is configured accordingly), handle_scheduling_exception comes in.
     """
-    # From here on, we handle IDs again, not objects
-    if isinstance(sensor, Sensor):
-        sensor = sensor.id
+    # We first create a scheduler and check if deserializing works, so the flex config is checked
+    # and errors are raised before the job is enqueued (so users get a meaningful response right away).
+    # Note: We are putting still serialized scheduler_kwargs into the job!
+    scheduler = find_scheduler_class(sensor)(sensor=sensor, **scheduler_kwargs)
+    scheduler.deserialize_config()
 
     job = Job.create(
         make_schedule,
-        kwargs=dict(sensor_id=sensor, **scheduler_kwargs),
+        kwargs=dict(sensor_id=sensor.id, **scheduler_kwargs),
         id=job_id,
         connection=current_app.queues["scheduling"].connection,
         ttl=int(
@@ -65,6 +64,7 @@ def create_scheduling_job(
     )
     if enqueue:
         current_app.queues["scheduling"].enqueue_job(job)
+
     return job
 
 
@@ -76,13 +76,16 @@ def make_schedule(
     belief_time: datetime | None = None,
     flex_model: dict | None = None,
     flex_context: dict | None = None,
+    flex_config_has_been_deserialized: bool = False,
 ) -> bool:
     """
-    This function is meant to be queued as a job. It returns True if it ran successfully.
+    This function computes a schedule. It returns True if it ran successfully.
 
-    Note: This function thus potentially runs on a different FlexMeasures node than where the job is created.
+    It can be queued as a job (see create_scheduling_job).
+    In that case, it will probably run on a different FlexMeasures node than where the job is created.
+    In any case, this function expects flex_model and flex_context to not have been deserialized yet.
 
-    This is what this function does
+    This is what this function does:
     - Find out which scheduler should be used & compute the schedule
     - Turn scheduled values into beliefs and save them to db
     """
@@ -112,6 +115,8 @@ def make_schedule(
         flex_model=flex_model,
         flex_context=flex_context,
     )
+    if flex_config_has_been_deserialized:
+        scheduler.config_deserialized = True
     consumption_schedule = scheduler.compute_schedule()
     if rq_job:
         click.echo("Job %s made schedule." % rq_job.id)
@@ -141,6 +146,8 @@ def make_schedule(
     ]  # For consumption schedules, positive values denote consumption. For the db, consumption is negative
     bdf = tb.BeliefsDataFrame(ts_value_schedule)
     save_to_db(bdf)
+
+    scheduler.persist_flex_model()
     db.session.commit()
 
     return True
@@ -207,7 +214,7 @@ def load_custom_scheduler(scheduler_specs: dict) -> type:
         try:
             module = importlib.import_module(module_descr)
         except TypeError as te:
-            current_app.log.error(f"Cannot load {module_descr}: {te}.")
+            current_app.logger.error(f"Cannot load {module_descr}: {te}.")
             raise
         except ModuleNotFoundError:
             current_app.logger.error(
@@ -234,7 +241,11 @@ def handle_scheduling_exception(job, exc_type, exc_value, traceback):
     """
     Store exception as job meta data.
     """
-    click.echo("HANDLING RQ WORKER EXCEPTION: %s:%s\n" % (exc_type, exc_value))
+    click.echo(
+        "HANDLING RQ SCHEDULING WORKER EXCEPTION: %s:%s\n" % (exc_type, exc_value)
+    )
+    # from traceback import print_tb
+    # print_tb(traceback)
     job.meta["exception"] = exc_value
     job.save_meta()