+++ documentation/api/change_log.rst
@@ -13,6 +13,7 @@ v3.0-5 | 2022-12-30
     - ``soc-at-start`` -> send in ``flex-model`` instead
     - ``soc-min`` -> send in ``flex-model`` instead
     - ``soc-max`` -> send in ``flex-model`` instead
+    - ``soc-targets`` -> send in ``flex-model`` instead
     - ``soc-unit`` -> send in ``flex-model`` instead
     - ``roundtrip-efficiency`` -> send in ``flex-model`` instead
     - ``prefer-charging-sooner`` -> send in ``flex-model`` instead
@@ -20,6 +21,8 @@ v3.0-5 | 2022-12-30
     - ``production-price-sensor`` -> send in ``flex-context`` instead
     - ``inflexible-device-sensors`` -> send in ``flex-context`` instead
 
+- Introduced the ``duration``field to `/sensors/<id>/schedules/trigger` (POST) for setting a planning horizon explicitly.
+- Allow posting ``soc-targets`` to `/sensors/<id>/schedules/trigger` (POST) that exceed the default planning horizon, and ignore posted targets that exceed the max planning horizon.
 - Added a subsection on deprecating and sunsetting to the Introduction section.
 - Added a subsection on describing flexibility to the Notation section.